Added k8

+++ src/ELibrary.Backend/UserApi/Command/Admin/AdminDeleteUser/AdminDeleteUserCommandHandler.cs
@@ -15,7 +15,10 @@ namespace UserApi.Command.Admin.AdminDeleteUser
 
         public async Task<Unit> Handle(AdminDeleteUserCommand command, CancellationToken cancellationToken)
         {
-            var user = await userService.GetUserByUserInfoAsync(command.Info, cancellationToken);
+            var user = await userService.GetUserByLoginAsync(command.Info, cancellationToken);
+
+            if (user == null) { throw new InvalidOperationException("User to delete is not found!"); }
+
             var result = await userService.DeleteUserAsync(user, cancellationToken);
             if (Utilities.HasErrors(result.Errors.ToList(), out var errorResponse)) throw new AuthorizationException(errorResponse);
             return Unit.Value;
